Update README.md
In this Kickstarter Challenge, I was to help an up and coming play writer named Louise help fund her play called Fever. In this data analysis, I was to help sort, organize, sort crowd funding data. This will help Louise get an idea on when to perform her funding raising tactics and on how much she may need to actually need to have as a goal to be  successful in creating her new play. I needed to determine what factors will be helpful to Louise in making campaign project successful.



In this file you will find the summary and conclusion of a kickstarter project